// src/models/schemas/RefreshToken.schema.ts
import { ObjectId } from 'mongodb'

interface RefreshTokenType {
  _id?: ObjectId
  token: string
  created_at?: Date
  user_id: ObjectId
  iat: number
  exp: number
}

export default class RefreshToken {
  _id?: ObjectId
  token: string
  created_at: Date
  user_id: ObjectId
  iat: Date
  exp: Date
  constructor({ _id, token, created_at, user_id, iat, exp }: RefreshTokenType) {
    this.token = token
    this.created_at = created_at || new Date()
    this.user_id = user_id
    this._id = _id
    this.iat = new Date(iat * 1000)
    this.exp = new Date(exp * 1000)
  }
}